refactor(compiler): move CFG utilities to Neo.Compiler.ControlFlow
The basic-block / control-flow utilities (BasicBlock, ContractInBasicBlocks,
InstructionCoverage, JumpTarget, OpCodeTypes, TryCatchFinallyCoverage) are
general control-flow-graph infrastructure shared by the optimizer and the
security analysers, not optimizer-private helpers. Move them from
Optimizer/Analysers (namespace Neo.Optimizer) into a dedicated ControlFlow
folder under namespace Neo.Compiler.ControlFlow, and update all references.

Inside the moved files, Instruction is aliased to Neo.VM.Instruction within
the namespace body so the enclosing Neo.Compiler scope does not shadow it
with the internal IR Instruction type.

Pure relocation; no behavioral change. Full compiler unit-test suite passes.
