+++ src/Neo.Compiler.CSharp/Optimizer/Strategies/JumpCompresser.cs
@@ -16,8 +16,8 @@ using Neo.VM;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using static Neo.Optimizer.JumpTarget;
-using static Neo.Optimizer.OpCodeTypes;
+using static Neo.Compiler.ControlFlow.JumpTarget;
+using static Neo.Compiler.ControlFlow.OpCodeTypes;
 
 namespace Neo.Optimizer
 {
