+++ src/Neo.Compiler.CSharp/SecurityAnalyzer/ReEntrancyAnalyzer.cs
@@ -11,6 +11,7 @@
 
 using Neo.Json;
 using Neo.Optimizer;
+using Neo.Compiler.ControlFlow;
 using Neo.SmartContract;
 using Neo.SmartContract.Manifest;
 using Neo.SmartContract.Native;
