+++ src/Neo.Compiler.CSharp/SecurityAnalyzer/TokenCallbackAuthorizationAnalyzer.cs
@@ -62,7 +62,7 @@ namespace Neo.Compiler.SecurityAnalyzer
                 if (instruction.OpCode != VM.OpCode.CALL && instruction.OpCode != VM.OpCode.CALL_L)
                     continue;
 
-                int target = Neo.Optimizer.JumpTarget.ComputeJumpTarget(addr, instruction);
+                int target = Neo.Compiler.ControlFlow.JumpTarget.ComputeJumpTarget(addr, instruction);
                 if (target >= 0)
                     methodStartOffsets.Add(target);
             }
@@ -132,7 +132,7 @@ namespace Neo.Compiler.SecurityAnalyzer
 
                     if (instruction.OpCode == VM.OpCode.CALL || instruction.OpCode == VM.OpCode.CALL_L)
                     {
-                        int target = Neo.Optimizer.JumpTarget.ComputeJumpTarget(addr, instruction);
+                        int target = Neo.Compiler.ControlFlow.JumpTarget.ComputeJumpTarget(addr, instruction);
                         if (methodStartOffsets.Contains(target))
                             pendingMethodStarts.Push(target);
                     }
