+++ src/Neo.Compiler.CSharp/SecurityAnalyzer/UnboundedOperationAnalyzer.cs
@@ -93,7 +93,7 @@ namespace Neo.Compiler.SecurityAnalyzer
                 if (instruction.OpCode != OpCode.CALL && instruction.OpCode != OpCode.CALL_L)
                     continue;
 
-                int target = Neo.Optimizer.JumpTarget.ComputeJumpTarget(addr, instruction);
+                int target = Neo.Compiler.ControlFlow.JumpTarget.ComputeJumpTarget(addr, instruction);
                 if (target >= 0)
                     methodStartOffsets.Add(target);
             }
@@ -104,7 +104,7 @@ namespace Neo.Compiler.SecurityAnalyzer
                 if (!instruction.OpCode.IsJumpInstruction())
                     continue;
 
-                int target = Neo.Optimizer.JumpTarget.ComputeJumpTarget(addr, instruction);
+                int target = Neo.Compiler.ControlFlow.JumpTarget.ComputeJumpTarget(addr, instruction);
                 if (target < addr)
                     backwardJumps.Add(addr);
             }
@@ -122,7 +122,7 @@ namespace Neo.Compiler.SecurityAnalyzer
                     if (instruction.OpCode != OpCode.CALL && instruction.OpCode != OpCode.CALL_L)
                         continue;
 
-                    int target = Neo.Optimizer.JumpTarget.ComputeJumpTarget(addr, instruction);
+                    int target = Neo.Compiler.ControlFlow.JumpTarget.ComputeJumpTarget(addr, instruction);
                     if (target == methodStart)
                         recursiveCalls.Add(addr);
                 }
