+++ tests/Neo.Compiler.CSharp.UnitTests/SecurityAnalyzer/UnitTest_Reentrancy.cs
@@ -13,6 +13,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Compiler.SecurityAnalyzer;
 using Neo.Json;
 using Neo.Optimizer;
+using Neo.Compiler.ControlFlow;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
 using Neo.SmartContract.Testing;
